HTGioiThieuViecLam/GioiThieuViecLamApp/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
migrate==0.3.7
packaging==23.2
Pillow==10.1.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
